fix(reservations): report the overlap when two agents reserve at once
Reading the conflicting leases and taking the lease were separate
statements, so two agents reserving the same path at once both read an
empty list before either wrote. Both took an exclusive lease and neither
was told about the other, which is the only thing an advisory lease is
for: it never blocks the edit, it warns while both agents can still talk
about it. Measured before the fix, both holders missed the clash in 16 of
50 concurrent attempts.

The read and the write now share one immediate transaction. The conflict
query runs on that transaction rather than the pooled handle, because the
store holds a single connection and a query issued on it while a
transaction is open waits forever.

Also:

- The refusal for a dash-leading path said it "reads as a flag", which
  this parser never does; both dash-leading and unknown-flag tokens are
  read as operands. The guard stays, since it catches a mistyped flag
  before it is leased as a pattern that can never conflict, but the
  sentence now says that.
- The CLI's feature tests move beside the files they cover, leaving the
  shared parser and help tests behind in cli_test.go.
- The terminal verbs join the table that asserts a missing id is a usage
  error, which covered only the session commands.
