+++ internal/cli/cli_test.go
@@ -192,141 +192,6 @@ func sampleSession() sessioncmd.Session {
 	}
 }
 
-func TestSessionCommandsParseArgumentsAndPrintSentences(t *testing.T) {
-	cases := []struct {
-		name    string
-		run     func(*bytes.Buffer, *fakeSessions, []string) error
-		args    []string
-		want    string
-		inspect func(*testing.T, *fakeSessions)
-	}{
-		{
-			name: "sessions lists every row",
-			run: func(out *bytes.Buffer, f *fakeSessions, args []string) error {
-				return runSessions(out, f, args, "cafe0001")
-			},
-			want: "- api-worker (beef1234) running claude in api at /repo; status=working; running=true",
-			inspect: func(t *testing.T, f *fakeSessions) {
-				if f.callerID != "cafe0001" {
-					t.Fatalf("caller = %q", f.callerID)
-				}
-			},
-		},
-		{
-			name: "send takes a target and a message",
-			run: func(out *bytes.Buffer, f *fakeSessions, args []string) error {
-				return runSend(out, f, args, "cafe0001")
-			},
-			args: []string{"beef1234", "ship it"},
-			want: "queued message 7 for session beef1234 at position 1",
-			inspect: func(t *testing.T, f *fakeSessions) {
-				if f.targetID != "beef1234" || f.message != "ship it" {
-					t.Fatalf("send got target %q message %q", f.targetID, f.message)
-				}
-			},
-		},
-		{
-			name: "read prints the pane",
-			run: func(out *bytes.Buffer, f *fakeSessions, args []string) error {
-				return runRead(out, f, args, "cafe0001")
-			},
-			args: []string{"beef1234"},
-			want: "screen text",
-		},
-		{
-			name: "kill names what it stopped",
-			run: func(out *bytes.Buffer, f *fakeSessions, args []string) error {
-				return runKill(out, f, args, "cafe0001")
-			},
-			args: []string{"beef1234"},
-			want: "killed api-worker (beef1234)",
-		},
-		{
-			name: "revive names what it brought back",
-			run: func(out *bytes.Buffer, f *fakeSessions, args []string) error {
-				return runRevive(out, f, args, "cafe0001")
-			},
-			args: []string{"beef1234"},
-			want: "revived api-worker (beef1234)",
-		},
-		{
-			name: "archive files a session away by default",
-			run: func(out *bytes.Buffer, f *fakeSessions, args []string) error {
-				return runArchive(out, f, args, "cafe0001")
-			},
-			args: []string{"beef1234"},
-			want: "archived api-worker (beef1234)",
-			inspect: func(t *testing.T, f *fakeSessions) {
-				if !f.archived {
-					t.Fatal("archive without --restore should archive")
-				}
-			},
-		},
-		{
-			name: "archive --restore puts it back",
-			run: func(out *bytes.Buffer, f *fakeSessions, args []string) error {
-				return runArchive(out, f, args, "cafe0001")
-			},
-			args: []string{"beef1234", "--restore"},
-			want: "restored api-worker (beef1234)",
-			inspect: func(t *testing.T, f *fakeSessions) {
-				if f.archived {
-					t.Fatal("--restore should unarchive")
-				}
-			},
-		},
-		{
-			name: "message-status reads the id back",
-			run: func(out *bytes.Buffer, f *fakeSessions, args []string) error {
-				return runMessageStatus(out, f, args, "cafe0001")
-			},
-			args: []string{"7"},
-			want: "message 7 to session beef is delivered",
-			inspect: func(t *testing.T, f *fakeSessions) {
-				if f.messageID != 7 {
-					t.Fatalf("message id = %d", f.messageID)
-				}
-			},
-		},
-		{
-			name: "groups lists the tree",
-			run: func(out *bytes.Buffer, f *fakeSessions, args []string) error {
-				return runGroups(out, f, args, "cafe0001")
-			},
-			want: "- api; sessions=2",
-		},
-		{
-			name: "create-group takes a path and a directory",
-			run: func(out *bytes.Buffer, f *fakeSessions, args []string) error {
-				return runCreateGroup(out, f, args, "cafe0001")
-			},
-			args: []string{"api/web", "--directory", "/repo"},
-			want: "created group api/web",
-			inspect: func(t *testing.T, f *fakeSessions) {
-				if f.groupPath != "api/web" || f.directory != "/repo" {
-					t.Fatalf("create-group got %q %q", f.groupPath, f.directory)
-				}
-			},
-		},
-	}
-
-	for _, testCase := range cases {
-		t.Run(testCase.name, func(t *testing.T) {
-			out := &bytes.Buffer{}
-			fake := &fakeSessions{session: sampleSession()}
-			if err := testCase.run(out, fake, testCase.args); err != nil {
-				t.Fatalf("run: %v", err)
-			}
-			if !strings.Contains(out.String(), testCase.want) {
-				t.Fatalf("output = %q, want it to contain %q", out.String(), testCase.want)
-			}
-			if testCase.inspect != nil {
-				testCase.inspect(t, fake)
-			}
-		})
-	}
-}
-
 // A shell caller reads the exit status and the error line, so a handler
 // that printed its sentence anyway would tell an agent the work happened.
 // The MCP front pins the same contract for its tools.
@@ -375,64 +240,6 @@ func TestALayerFailureReachesTheCaller(t *testing.T) {
 				t.Fatalf("a failed call printed a result anyway: %q", out.String())
 			}
 		})
-	}
-}
-
-// Only a flag the caller typed may reach the layer: a zero value passed
-// anyway would silently override the group and worktree a spawn inherits.
-func TestSpawnPassesOnlyTheFlagsGiven(t *testing.T) {
-	out := &bytes.Buffer{}
-	fake := &fakeSessions{session: sampleSession()}
-	args := []string{"--name", "api-worker", "--prompt", "build the api", "--tool", "claude", "--directory", "/repo"}
-	if err := runSpawn(out, fake, args, "cafe0001"); err != nil {
-		t.Fatalf("spawn: %v", err)
-	}
-	if fake.opts.Name != "api-worker" || fake.opts.Prompt != "build the api" || fake.opts.Tool != "claude" || fake.opts.Directory != "/repo" {
-		t.Fatalf("spawn opts = %+v", fake.opts)
-	}
-	if fake.opts.Group != nil || fake.opts.Worktree != nil {
-		t.Fatalf("untyped flags should stay inherited, got group=%v worktree=%v", fake.opts.Group, fake.opts.Worktree)
-	}
-	if !strings.HasPrefix(out.String(), "created api-worker (beef1234)") {
-		t.Fatalf("spawn output = %q", out.String())
-	}
-
-	explicit := &fakeSessions{session: sampleSession()}
-	if err := runSpawn(&bytes.Buffer{}, explicit, []string{"--group", "", "--worktree"}, "cafe0001"); err != nil {
-		t.Fatalf("spawn explicit: %v", err)
-	}
-	if explicit.opts.Group == nil || *explicit.opts.Group != "" {
-		t.Fatalf("an explicit empty group targets the root, got %v", explicit.opts.Group)
-	}
-	if explicit.opts.Worktree == nil || !*explicit.opts.Worktree {
-		t.Fatalf("--worktree should be passed on, got %v", explicit.opts.Worktree)
-	}
-}
-
-func TestWaitCollectsStatesAndFailsOnTimeout(t *testing.T) {
-	reached := &fakeSessions{wait: sessioncmd.WaitResult{
-		Session: sampleSession(), Reached: true, Outcome: sessioncmd.WaitReached, Waited: "3s", ManagerAwake: true,
-	}}
-	out := &bytes.Buffer{}
-	if err := runWait(out, reached, []string{"beef1234", "--until", "idle,finished", "--timeout", "10s"}, "cafe0001"); err != nil {
-		t.Fatalf("wait: %v", err)
-	}
-	if len(reached.until) != 2 || reached.until[0] != "idle" || reached.until[1] != "finished" {
-		t.Fatalf("until = %v", reached.until)
-	}
-	if reached.timeout != 10*time.Second {
-		t.Fatalf("timeout = %s", reached.timeout)
-	}
-	if !strings.Contains(out.String(), "is working after 3s") {
-		t.Fatalf("wait output = %q", out.String())
-	}
-
-	timedOut := &fakeSessions{wait: sessioncmd.WaitResult{
-		Session: sampleSession(), Outcome: sessioncmd.WaitTimedOut, Waited: "50s", ManagerAwake: true,
-	}}
-	err := runWait(&bytes.Buffer{}, timedOut, []string{"beef1234"}, "cafe0001")
-	if err == nil || !strings.HasPrefix(err.Error(), "timed out:") {
-		t.Fatalf("a timeout should fail the command, got %v", err)
 	}
 }
 
@@ -500,13 +307,6 @@ func TestMissingSessionIDIsAUsageError(t *testing.T) {
 	}
 }
 
-func TestMessageStatusRefusesANonNumericID(t *testing.T) {
-	err := runMessageStatus(&bytes.Buffer{}, &fakeSessions{}, []string{"seven"}, "cafe0001")
-	if err == nil || !strings.Contains(err.Error(), "agent-manager send prints the id it queued") {
-		t.Fatalf("error = %v, want it to say where the id comes from", err)
-	}
-}
-
 func TestHelpFlagPrintsUsageAndStops(t *testing.T) {
 	out := &bytes.Buffer{}
 	fake := &fakeSessions{session: sampleSession()}
@@ -530,64 +330,6 @@ func TestHelpFlagPrintsUsageAndStops(t *testing.T) {
 	}
 }
 
-func TestTaskVerbsDispatch(t *testing.T) {
-	out := &bytes.Buffer{}
-	fake := &fakeSessions{task: sessioncmd.Task{ID: "t1", Title: "wire the cli", State: "in_progress", OwnerName: "api-worker"}}
-
-	if err := runTaskList(out, fake, nil, "cafe0001"); err != nil {
-		t.Fatalf("task list: %v", err)
-	}
-	if !strings.Contains(out.String(), "- wire the cli (t1) [in_progress] held by api-worker") {
-		t.Fatalf("task list output = %q", out.String())
-	}
-
-	created := &bytes.Buffer{}
-	if err := runTaskCreate(created, fake, []string{"wire the cli", "--body", "five files", "--depends-on", "a1,b2"}, "cafe0001"); err != nil {
-		t.Fatalf("task create: %v", err)
-	}
-	if fake.title != "wire the cli" || fake.body != "five files" {
-		t.Fatalf("task create got %q / %q", fake.title, fake.body)
-	}
-	if len(fake.dependsOn) != 2 || fake.dependsOn[0] != "a1" || fake.dependsOn[1] != "b2" {
-		t.Fatalf("depends-on = %v", fake.dependsOn)
-	}
-
-	if err := runTaskClaim(&bytes.Buffer{}, fake, nil, "cafe0001"); err != nil {
-		t.Fatalf("task claim: %v", err)
-	}
-	if fake.taskID != "" {
-		t.Fatalf("a bare claim takes the next task, got id %q", fake.taskID)
-	}
-	if err := runTaskClaim(&bytes.Buffer{}, fake, []string{"t1"}, "cafe0001"); err != nil {
-		t.Fatalf("task claim t1: %v", err)
-	}
-	if fake.taskID != "t1" {
-		t.Fatalf("claim id = %q", fake.taskID)
-	}
-
-	settled := map[string]func(io.Writer, taskCommands, []string, string) error{
-		"finished": runTaskFinish,
-		"released": runTaskRelease,
-	}
-	for verb, run := range settled {
-		buf := &bytes.Buffer{}
-		if err := run(buf, fake, []string{"t1"}, "cafe0001"); err != nil {
-			t.Fatalf("task %s: %v", verb, err)
-		}
-		if !strings.HasPrefix(buf.String(), verb+" wire the cli (t1)") {
-			t.Fatalf("task %s output = %q", verb, buf.String())
-		}
-	}
-
-	deleted := &bytes.Buffer{}
-	if err := runTaskDelete(deleted, fake, []string{"t1"}, "cafe0001"); err != nil {
-		t.Fatalf("task delete: %v", err)
-	}
-	if !fake.deleted || deleted.String() != "deleted task t1\n" {
-		t.Fatalf("task delete output = %q", deleted.String())
-	}
-}
-
 func TestTaskGroupRejectsAnUnknownVerb(t *testing.T) {
 	err := dispatch(&bytes.Buffer{}, "task", taskVerbs(), []string{"finnish", "t1"}, "cafe0001", t.TempDir())
 	if err == nil {
@@ -601,114 +343,6 @@ func TestTaskGroupRejectsAnUnknownVerb(t *testing.T) {
 	bare := dispatch(&bytes.Buffer{}, "task", taskVerbs(), nil, "cafe0001", t.TempDir())
 	if bare == nil || bare.Error() != "usage: agent-manager task <list|create|claim|finish|release|delete>" {
 		t.Fatalf("a bare group should print its verbs, got %v", bare)
-	}
-}
-
-func TestReserveParsesPathsAndFlags(t *testing.T) {
-	fake := &fakeSessions{reserve: sessioncmd.ReserveResult{
-		Reserved: []sessioncmd.Reservation{{Pattern: "internal/cli", Mode: "exclusive"}},
-	}}
-	out := &bytes.Buffer{}
-	args := []string{"internal/cli", "internal/ui", "--mode", "shared", "--note", "wiring", "--ttl", "45m"}
-	if err := runReserve(out, fake, args, "cafe0001"); err != nil {
-		t.Fatalf("reserve: %v", err)
-	}
-	if len(fake.patterns) != 2 || fake.patterns[1] != "internal/ui" {
-		t.Fatalf("patterns = %v", fake.patterns)
-	}
-	if fake.mode != "shared" || fake.note != "wiring" || fake.ttl != 45*time.Minute {
-		t.Fatalf("reserve flags = %q %q %s", fake.mode, fake.note, fake.ttl)
-	}
-	if !strings.Contains(out.String(), "reserved internal/cli") {
-		t.Fatalf("reserve output = %q", out.String())
-	}
-
-	if err := runReserve(&bytes.Buffer{}, fake, []string{"--", "-notaflag"}, "cafe0001"); err == nil {
-		t.Fatal("a path that reads as a flag should be refused")
-	}
-
-	// The flags parse wherever they sit, so a leading dash is the whole
-	// reason a mistyped one cannot be leased, and the refusal has to say so
-	// rather than send the agent to reorder what it typed.
-	mistyped := runReserve(&bytes.Buffer{}, fake, []string{"internal/cli", "--mod", "shared"}, "cafe0001")
-	if mistyped == nil {
-		t.Fatal("a mistyped flag should not be leased as a pattern")
-	}
-	if !strings.Contains(mistyped.Error(), `"--mod" starts with a dash`) ||
-		!strings.Contains(mistyped.Error(), "usage: agent-manager reserve") {
-		t.Fatalf("the refusal does not name the cause and the usage: %q", mistyped)
-	}
-	if err := runReserve(&bytes.Buffer{}, fake, nil, "cafe0001"); err == nil {
-		t.Fatal("reserve needs at least one path")
-	}
-}
-
-func TestReservationsAndReleaseReadTheSameLeases(t *testing.T) {
-	out := &bytes.Buffer{}
-	if err := runReservations(out, &fakeSessions{}, nil, "cafe0001"); err != nil {
-		t.Fatalf("reservations: %v", err)
-	}
-	if !strings.Contains(out.String(), "- internal/cli (exclusive) held by api for 30m0s") {
-		t.Fatalf("reservations output = %q", out.String())
-	}
-
-	released := &bytes.Buffer{}
-	fake := &fakeSessions{released: 2}
-	if err := runReleaseFiles(released, fake, nil, "cafe0001"); err != nil {
-		t.Fatalf("release-files: %v", err)
-	}
-	if len(fake.patterns) != 0 {
-		t.Fatalf("naming no path releases everything, got %v", fake.patterns)
-	}
-	if released.String() != "released 2 reservation(s)\n" {
-		t.Fatalf("release-files output = %q", released.String())
-	}
-}
-
-func TestTerminalVerbsDispatch(t *testing.T) {
-	fake := &fakeTerminals{}
-	out := &bytes.Buffer{}
-	if err := runTerminalList(out, fake, nil, "cafe0001"); err != nil {
-		t.Fatalf("terminal list: %v", err)
-	}
-	if !strings.Contains(out.String(), "- build (t1) in root at /repo; status=idle; running=true") {
-		t.Fatalf("terminal list output = %q", out.String())
-	}
-
-	created := &bytes.Buffer{}
-	if err := runTerminalCreate(created, fake, []string{"--directory", "/repo"}, "cafe0001"); err != nil {
-		t.Fatalf("terminal create: %v", err)
-	}
-	if fake.opts.Directory != "/repo" || fake.opts.Group != nil {
-		t.Fatalf("terminal create opts = %+v", fake.opts)
-	}
-
-	sent := &bytes.Buffer{}
-	if err := runTerminalSend(sent, fake, []string{"t1", "--command", "go test ./..."}, "cafe0001"); err != nil {
-		t.Fatalf("terminal send: %v", err)
-	}
-	if fake.command != "go test ./..." || sent.String() != "sent command to terminal t1\n" {
-		t.Fatalf("terminal send got %q, printed %q", fake.command, sent.String())
-	}
-
-	keyed := &bytes.Buffer{}
-	if err := runTerminalSend(keyed, fake, []string{"t1", "--keys", "C-c,Up"}, "cafe0001"); err != nil {
-		t.Fatalf("terminal send keys: %v", err)
-	}
-	if len(fake.keys) != 2 || fake.keys[0] != "C-c" || keyed.String() != "sent keys to terminal t1\n" {
-		t.Fatalf("terminal send keys = %v, printed %q", fake.keys, keyed.String())
-	}
-
-	read := &bytes.Buffer{}
-	if err := runTerminalRead(read, fake, []string{"t1"}, "cafe0001"); err != nil {
-		t.Fatalf("terminal read: %v", err)
-	}
-	if read.String() != "build passed\n" {
-		t.Fatalf("terminal read output = %q", read.String())
-	}
-
-	if err := dispatch(&bytes.Buffer{}, "terminal", terminalVerbs(), []string{"tail"}, "cafe0001", t.TempDir()); err == nil {
-		t.Fatal("an unknown terminal verb should not dispatch")
 	}
 }
